refactor: Consolidate account routes, use ts extension.
